SF: Clean up updateTransformHint() processing.
While working on ag/3368935, an unintentional duplicate call was added
to Layer::updateTransform() from
SurfaceFlinger::handleTransactionLocked(). This patch removes the
duplicate call.

In examining the code, the update conditions also did not appear to be
set correctly. They have been adjusted so that a display change or a
significant layer change (such as a layer being added) will result in an
update.

Bug: 76223411
Test: No immediate issues observed on a Pixel XL.

Change-Id: I59220837b22ae1f76d7d9d376c1ebd24d8d14105