Do not allocate protected AHardwareBuffers with CPU access.
Depending on the gralloc implementation, allocating AHardwareBuffers
with PROTECTED_CONTENT usage and nonzero CPU read or write mask may
succeed, but the buffer won't be accessible by the CPU - either
because the lock() call will fail, or there will be memory violation
when trying to access the memory. Prevent allocating such buffers.

Bug: 77461051
Test: Builds and passes CtsNativeHardwareTestCases on Pixel XL.
Change-Id: I822c9fb2d8ce24cd0c0fc0ac765b7a71fd372199