Expose EGL_EXT_image_gl_colorspace if EGL_KHR... is present.
Adreno 530 driver exposes EGL_KHR_image_gl_colorspace, which is
the draft name of EGL_EXT_image_gl_colorspace that was changed during
Khronos review. Add a mitigation to expose the correct extension
string until the driver can be fixed.

Bug: 78368937
Test: Built and flashed Pixel XL, verified that the extension
EGL_EXT_image_gl_colorspace is present in the list of EGL exts;
Ran CtsNativeHardwareTestCases and inspected the log to verify
that test cases for sRGB EGLImages are not skipped.

Change-Id: I64cef7be828516e2dbc160f98c83879fd8662450