Fix eglMakeCurrent crash when in opengl contexts
Nvidia shield devices support using egl to switch to a full desktop
opengl context. In opengl 3.0+, GL_EXTENSIONS have to be retrieved
with glGetStringi and is invalid for glGetString. Thus,
eglMakecurrent can crash with a NPE if this case is not handled.
The logic here is building a wrapper for glGetStringi, thus the
error can be ignored.

Change-Id: I9c599e10c62aabf684bde4e81719aa248327ac80