Input tests no longer use test_per_src
Do not specify test_per_src = true for input native tests. Therefore,
only a single binary will be generated for each test suite.

Input tests, for historical reasons, used a variable 'test_per_src' in
the makefile, meaning that a new make target will be generated for each
test C++ file. However, since the tests are using gtest, this is not
necessary. This was the case since the early days of transition from .mk
to .bp. At this point, input system is the only gtest suite that uses
this flag, and there's no merit to continue using it or maintaining.
The added benefit is that now, atest will automatically start supporting
the input native tests.

Bug: none
Test: atest libinput_tests inputflinger_tests
Change-Id: Ie762d31f0b5be5f608dd84bd86dffc566e6706bb