[SurfaceFlinger] Hide SyncFeatures behind RenderEngine
This allows a future test that uses BufferLayerConsumer as-is and does
not actually set up a GL context to work by adding call expectations to
RenderEngine for what sync configuration should be used.

SyncFeatures is a singleton, and makes a bunch of GL queries when first
constructed to determine what is supported by the GL driver. It made
most sense to me to hide this inside the RenderEngine for that reason.

Test: Builds
Bug: None

Change-Id: Ic62bfb689a84fd9e509139c7f189223cf215277e