Fix landscape LCM issue
When device uses landscape LCM, the nature of the screen shows that
the device is a portrait device. In this case, we find GPU filter
function is opened by mistake when capturing screen and CTS case
fails because image pixel value is changed.

The solution is to correct the GPU filter by hworientation.

Bug: 69691076
Test:
1. Capture screen
2. android.uirendering.cts.testclasses.SurfaceViewTests#testMovingWhiteSurfaceView

* Original patch submitted at:
  https://android-review.googlesource.com/c/platform/frameworks/native/+/577262

* Also backported for the libsurfaceflinger implementation that uses
  legacy libhwcomposer (original patch is for the hwc2-compatible
  libsurfaceflinger only).

Change-Id: I67b75fc8fab188d24f1d5febff20d9bd10c15204
Signed-off-by: Vladimir Oltean <olteanv@gmail.com>