Fix display power state reset in vr flinger path
Some code refactoring in ag/4355343 accidentally broke vr flinger. When
switching to vr flinger, we call setPowerModeInternal() to reset the
display state. With ag/4355343, we were incorrectly holding on to a
stale display, causing us to early exit from setPowerModeInternal() and
not apply the display power state change.

Bug: 111398862

Test: Manually ran vr flinger, confirmed it works again after applying
this CL.

Change-Id: Idbead0338f9219c80cb8a2d9f283daefff373a33