Changed getViewport strategy
If a uniqueId is specified, then it must match. If it is not specified,
then return the first display view port that has the requested type.
This means that if an internal or external viewport does not have any
uniqueId specified, there should only be 1 of each of these. Otherwise,
there will be confusion, and the viewports are not kept in any specific
order.

Test: atest -a libinput_tests inputflinger_tests
Bug: 111108021
Change-Id: I2437104ad98fa24334f2a8e19f1ca330bd873c93