SF: Enable address sanitizer for the unittest
Not only will it catch issues in the code covered by the test, but it
will also catch issues in how the test injects fakes and similar objects
into the code under test.

In terms of runtime costs, it makes the tests about 2x slower, so now it
takes 200ms to run instead of 100ms.

Bug: None
Test: atest libsurfaceflinger_unittest

Change-Id: I8de3e65c4e7e16cef9f03ed0ba038216971b957d