Use opt-in/out logic that lives in the ANGLE APK
The code dlopen's the ANGLE version of libGLESv2.so and then
dlsym/call the ANGLEUseForApplication() function that lives in that
library.  It only calls the opt-in/out logic if the Developer Options
doesn't force the use of ANGLE.  If ANGLE can't be loaded, it also
goes with the application's preference.

Test: Run applications and look at log messages.
Change-Id: I7fa8dec991fbc217157dc12e6466ed4df5fb4020
(cherry picked from commit 85ce86539ac6d56e1c2a37eabed930be82ab4439)