Remove minLayer and maxLayer from screenshot code.
When taking a full display screenshot, there's no longer a need for min
and max layer. Instead, the entire display will be screenshotted. If
only some layers should be screenshotted, then use the captureLayers method
instead. This was already the case for real code, but required some
clean up in Transaction_test

Test: Transaction_test
Change-Id: I774357b284ca9e026097f79fb7146e4931fa629a