surfaceflinger: use std::vector for mDisplayEventConnections
Replace SortedVector by std::vector for mDisplayEventConnections.
There are usually a couple dozens of connections.  They are created
and destroyed as app processes come and go, and are iterated
frequently.

Semantically, SortedVector is std::set rather than std::vector.  But
registerDisplayEventConnection can be made private, and the sole
caller never adds the same connection twice, we choose to replace it
by std::vector.  However, we still keep a check in
registerDisplayEventConnection to make sure it is indeed the case
temporarily.

Bug: 115738279
Test: boots
Change-Id: Ie2df1f346cceda272f58f7751b31534171d593ca