OtaDexopt: Temporarily disable
Exit-early from the OTA script to avoid a bug. We need to skip an
OTA as the code in question runs in the host part.

It is easier to early-return than to remove to postinstall hook,
as that is registered on a per-product basis.

Bug: 115853663
Test: m
Change-Id: I8ebcdd7d800c5c4d27a0f07fb4145b88fd19791e
(cherry picked from commit c0ba0e4b0882a2454f162c453099d4371fc66b15)