surfaceflinger: free SurfaceTracing memory
Allocate LayersTraceFileProto on the heap because Clear does not
free the memory.  Replace mTrace->Clear() by mTrace.reset().

This is temporary.  Eventually, we want to enable continuous tracing
and dump only the last N frames.  We will record LayersProto in a
circular buffer and create LayersTraceFileProto only to serialize
the buffer to a file.

Bug: 115434782
Test: winscope trace
Change-Id: I3a3e499ace08cd4b06c38be8ad2ad59bfcdba5d7