[SF] Fix unittest crash
(Manual cherrypick to aosp)

Some of the values passed to HWC2::Display by FakeHwcDisplayInjector as
used by the test were passed by reference, with the reference being
owned by the injector.

However, the lifetime of the injector was effectively temporary, and so
the constructed Display instance ended up holding a reference to a dead
instance.

This patch uses statically constructed default values for those
parameters instead, so that the values will not be destroyed. The test
code can also still override the values used if non-default values are
ever desired.

It also adds a comment warning of the possibility so it hopefully does
not happen again.

Bug: 111454533
Bug: 117451794
Test: atest libsurfaceflinger_unittest # On aosp_cf_x86_phone-userdebug

Change-Id: I8dfe6e31111a558a7892f0eef91bd96b8dd71f59
Merged-In: I8dfe6e31111a558a7892f0eef91bd96b8dd71f59