Add createBuffer to BufferHubBinderService
When calling createBuffer, bufferhubd will now alloc the buffer via
creating a BufferNode in the server side, and return you a
BpBufferClient object for communication. A BpBufferClient is binded with
one specific buffer.

Currently you could only call isValid() on the client. More APIs will be
added in future CL.

Test: "atest buffer_hub_binder_service-test". Passed
Bug: 116681016
Change-Id: I05ec627474303af46a792b0b6c2eaa904724d1f2