Flag invalid robust access attribute
The vendor driver may support EGL 1.5 while only EGL 1.4 is presented
to applications. According to the specification,
EGL_CONTEXT_OPENGL_RESET_NOTIFICATION_STRATEGY is not supported with
GLES until EGL 1.5. Because the vendor driver may be a different
version than what's presented to applications, throw an error on the invalid
attribute rather than depend on the vendor driver to flag the error.
bug: 111083885
Test: dEQP-EGL.functional.robustness.negative_context.invalid_notification_strategy_enum

Change-Id: I22702f74700d834667bed92939fcabab5f0726e8